Add qmake file for building a shared library
