+++ .gitignore
@@ -2,6 +2,7 @@
 ui_*.h
 *.o
 *.so
+*.so.*
 moc_*.cpp
 Makefile
 out/*
