+++ include/Color_Delegate
@@ -1,0 +1,1 @@
+#include "color_delegate.hpp"
