+++ include/Color_Dialog
@@ -1,0 +1,1 @@
+#include "color_dialog.hpp"
