+++ include/bound_color_selector.hpp
@@ -24,40 +24,24 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 */
 
-#ifndef COLOR_LIST_WIDGET_HPP
-#define COLOR_LIST_WIDGET_HPP
+#ifndef BOUND_COLOR_SELECTOR_HPP
+#define BOUND_COLOR_SELECTOR_HPP
 
-#include "abstract_widget_list.hpp"
+#include "color_selector.hpp"
 
-class Color_List_Widget : public Abstract_Widget_List
+/**
+ *  A color selector bound to a color reference
+ */
+class QCP_EXPORT Bound_Color_Selector : public Color_Selector
 {
     Q_OBJECT
-
-    Q_PROPERTY(QList<QColor> colors READ colors WRITE setColors NOTIFY colorsChanged )
-
+private:
+    QColor* ref;
 public:
-    explicit Color_List_Widget(QWidget *parent = 0);
-    ~Color_List_Widget();
-
-    QList<QColor> colors() const;
-    void setColors(const QList<QColor>& colors);
-
-    void swap(int a, int b);
-
-    void append();
-
-signals:
-    void colorsChanged(const QList<QColor>&);
+    explicit Bound_Color_Selector(QColor* reference, QWidget *parent = 0);
 
 private slots:
-    void emit_changed();
-    void handle_removed(int);
-    void color_changed(int row);
-
-private:
-    class Private;
-    Private * const p;
-    void  append_widget(int col);
+    void update_reference(QColor);
 };
 
-#endif // COLOR_LIST_WIDGET_HPP
+#endif // BOUND_COLOR_SELECTOR_HPP
